loader: fold reference_test.go into include_test.go
Both reference tests cover include / env_file scoping and now belong
next to the rest of the include tests rather than in their own file.
Drops the now-empty reference_test.go.
